"[Fix] Sns, Like, BookMark Service 타입에 맞게 수정"

+++ src/main/java/com/jobdam/sns/repository/SnsPostRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface SnsPostRepository extends JpaRepository<SnsPost, Long> {
+public interface SnsPostRepository extends JpaRepository<SnsPost, Integer> {
 
     List<SnsPost> findByUserId(Integer userId);
     List<SnsPost> findByTitleContaining(String keyword);
